requirements.txt
spacy~=3.7.2
torch~=2.1.2
transformers~=4.31.0
requests~=2.31.0
streamlit~=1.29.0
streamlit-authenticator~=0.2.3
pandas~=1.5.3
pydantic~=2.5.2
numpy~=1.26.2
beautifulsoup4~=4.12.2
selenium~=4.16.0
textacy~=0.13.0
scikit-learn~=1.3.2
kor~=1.0.0
langchain~=0.0.351
pypdf2~=2.10.5
networkx~=3.2.1
configparser~=5.0.2
python-dotenv~=1.0.0
scipy~=1.11.4
validators~=0.18.2
pyseoanalyzer~=4.0.7
plotly==5.18.0
plotly-express==0.4.1
SQLAlchemy~=2.0.23
DateTime
pycountry~=23.12.11
cleanco~=2.2
python-Levenshtein~=0.23.0
python-lei~=0.1.1


